backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.1.2
gunicorn==21.2.0
dnspython==2.4.2
